+++ dashboard/apps/prototype/static/prototype/js/project.js
@@ -4,6 +4,7 @@ import 'whatwg-fetch';
 import URI from 'urijs';
 import moment from 'moment';
 import _ from 'lodash';
+import Plotly from './custom-plotly';
 
 /**
  * send a POST request to the backend to retrieve project profile
@@ -64,4 +65,81 @@ export function getProjectId() {
 export function loadProjectPage(id) {
   const url = [location.protocol, '//', location.host, location.pathname].join('');
   window.location.href = url + '?projectid=' + id;
+}
+
+
+/**
+ * plot the graphs for a project
+ */
+export function plotProject(project, elem) {
+  const financial = parseProjectFinancials(project.financial);
+  const months = financial.months;
+
+  const civilServiceTrace = {
+    x: months,
+    y: financial.civilServantCosts,
+    name: 'Civil Servant',
+    type: 'bar',
+    marker: {
+      color: '#c0c2dc'
+    }
+  };
+  const contractorTrace = {
+    x: months,
+    y: financial.contractorCosts,
+    name: 'Contractor',
+    type: 'bar',
+    marker: {
+      color: '#b5d8df'
+    }
+  };
+  const additionalTrace = {
+    x: months,
+    y: financial.additionalCosts,
+    name: 'additional',
+    type: 'bar',
+    marker: {
+      color: '#B2CFB2'
+    }
+  };
+  const totalCostTrace = {
+    x: months,
+    y: financial.totalCostsCumulative,
+    name: 'Cumulative',
+    mode: 'lines',
+    yaxis: 'y2',
+    marker: {
+      color: '#6F777B'
+    }
+  };
+  const budgetTrace = {
+    x: months,
+    y: financial.budget,
+    name: 'Budget',
+    mode: 'lines',
+    yaxis: 'y2',
+    marker: {
+      color: '#F0E891'
+    }
+  };
+  const layout = {
+    title: project.name,
+    barmode: 'stack',
+    yaxis: {
+      title: 'monthly cost'
+    },
+    yaxis2: {
+      title: 'cumulative',
+      overlaying: 'y',
+      side: 'right'
+    },
+    legend: {
+      yanchor: 'bottom'
+    }
+  };
+  Plotly.newPlot(
+      elem,
+      [civilServiceTrace, contractorTrace, additionalTrace, totalCostTrace,
+       budgetTrace],
+      layout);
 }
